Add icons to Actions when choosing for widget/shortcuts
New iOS 14 API lets us assign an image to show up along with the actions.
